requirements.txt
streamlit
requests
python-dotenv
openai==1.30.5
agents
git+https://github.com/nestordemeure/agents.git
